requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
WTForms==3.1.2
Flask-WTF==1.2.1
Bootstrap-Flask==2.4.0
Werkzeug==3.0.3
email_validator==2.1.1